Revert "Break AnnotationNode into its own file so I don't have 5 minutes of searching around every time I look at the class hierarchy."
This reverts commit 24fdf6c93980e2b3952c312aea47f0d35fec8a4c.
